OnTrackEnd and OnTrackLoop added
Implemented OnTrackEnd and OnTrackLoop events. Wrote unit tests for
them, but one of them is flaky for very quickly-looping songs. This is
likely an inherent problem with Unity's InvokeRepeating function and is
probably unlikely to be fixed.

One of the other unit tests is failing. Committing to create a
checkpoint.