Fixed crossfading
Crossfading is fixed! Now everything works--just not the same as before.

Going forward, all fade functions do not mutate the volume setting on the player itself. Attempts to do so will invariably introduce code that breaks how crossfading works, as that uses the same fading mechanisms.

Perhaps in the future this will change so that publically-facing fade functions will also mutate the player's volume, but it is not today.